♻️ [Refactoring]: #455 Phase 0 literal_string テストモジュールを子モジュールへ退避
L199-835 の `mod tests { ... }` の中身を `literal_string/tests.rs` に
外側ラッパ除去 + 一律 unindent のみで退避し、本体側を
`#[cfg(test)] mod tests;` の 1 行宣言に圧縮。テスト関数 73 件の
名前・本体・直下日本語コメントは完全保持。
